Make Jenkins call the SQ URL defined in the SonarInstallation, instead of SQ public URL.
Make Jenkins call the SQ URL defined in the SonarInstallation, instead of SQ public URL.
To make this work with the corner-case requirements from SONARJNKNS-287 (server
URL defined via some variables), we store the expanded URL in SonarAnalysisAction.

Note that vars expansion for the URL is not implemented in SonarPublisher/SonarMaven
(would be possible, but it's deprecated code, so we keep its old behavior unchanged).