Use SQ URL from SonarInstallation, not public URL
Make Jenkins call the SQ URL defined in the SonarInstallation, instead of SQ public URL.
To make this work with the corner-case requirements from SONARJNKNS-287 (SonarInstallation
URL defined via some variables), we store the var-expanded URL in SonarAnalysisAction, and
reuse this URL in waitForQualityGate.  If the SonarAnalysisAction has been stored without
a ready-to-use server URL (for instance if it contains variables we could not expand, then
the SQ public URL (from report-task.txt) will still be used as a fallback.