Add a timeout option to the HttpClient to fix a bug
Naive first pass at introducing a default timeout
to HttpClient.java to fix a bug where the request will hang
on a non-response from a call to /api/server/version from
the Jenkins Master. The default timeout from HttpConnector is
a very long 30s to connect and 60s to read.