Fixed some compilation issues and removed arg for timeout
I got my IDE to function properly and fixed some errors that
the linter picked up. I also realize that as far as I have
discovered in Java I can't set an optional arg for timeout_ms
like in Python so I would either need to refactor all instances
of getHttp to include null for that third timeout_ms arg, or
just set defaults for now to get this to build.